requirements.txt
# PulseBMS Enhanced - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# MQTT & Communication
paho-mqtt==1.6.1
websockets==12.0
python-socketio==5.10.0

# Battery Modeling & Physics
pybamm==23.9
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Machine Learning & RL
torch==2.1.1
stable-baselines3==2.2.1
gymnasium==0.29.1
tensorboard==2.15.1

# Optimization
cvxpy==1.4.1
casadi==3.6.4  # For MPC
gekko==1.0.6   # Alternative optimization

# Data Processing
pandas==2.1.4
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing FastAPI

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
